+++ app/services/insights/generate_service.rb
@@ -39,7 +39,12 @@ module Insights
     def insight_attributes(entity_id)
       insight_fields.inject({}) do |acc, insight_field|
         unless insight_field.ends_with?('ratio')
-          next acc.merge({ insight_field.to_sym => send(insight_field)[entity_id].to_i })
+          value = send(insight_field)[entity_id]
+          decimal = Insight::DECIMAL_ATTRIBUTES.include?(insight_field.to_sym)
+
+          next acc.merge({
+            insight_field.to_sym => (decimal ? value.to_f : value.to_i)
+          })
         end
 
         acc.merge({ insight_field.to_sym => send(:ratio, insight_field[0..-7], entity_id) })